Change all instances of size_t to int to avoid silent overflow bugs and compress integer types.

+++ cpp/algorithm_nll.hpp
@@ -5,27 +5,27 @@
 #include "sampling.hpp"
 
 namespace algorithm{
-    std::vector<sample_data> load_data_from_file(std::string&,size_t&,size_t&,size_t&);
-    std::vector<size_t> load_data_labels_from_file(std::string&,size_t&,size_t&);
+    std::vector<sample_data> load_data_from_file(std::string&,int&,int&,int&);
+    std::vector<int> load_data_labels_from_file(std::string&,int&,int&);
     //nnTTN functions
     template<typename cmp>
-    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<size_t>&,std::vector<sample_data>&,std::vector<size_t>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<int>&,std::vector<sample_data>&,std::vector<int>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<size_t>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<int>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<sample_data>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll(graph<cmp>&,std::vector<sample_data>&,std::vector<sample_data>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll(graph<cmp>&,std::vector<sample_data>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll(graph<cmp>&,std::vector<sample_data>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,int>&,bool);
     
     //TTNBM functions
     template<typename cmp>
-    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<size_t>&,std::vector<sample_data>&,std::vector<size_t>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<int>&,std::vector<sample_data>&,std::vector<int>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<size_t>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<int>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<sample_data>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,std::vector<sample_data>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,double>&,std::map<int,int>&,bool);
     template<typename cmp>
-    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,size_t,size_t,bool,double,size_t,std::map<size_t,double>&,std::map<size_t,size_t>&,bool);
+    void train_nll_born(graph<cmp>&,std::vector<sample_data>&,int,int,bool,double,int,std::map<int,double>&,std::map<int,int>&,bool);
     
     template<typename cmp>
     void calculate_site_probs(graph<cmp>&,bond&);
