+++ cpp/sampling.hpp
@@ -13,21 +13,21 @@
 class sample_data{
 public:
     sample_data();
-    sample_data(size_t,std::vector<size_t>);
-    size_t n_phys_sites() const;
-    std::vector<size_t> s() const;
-    size_t& n_phys_sites();
-    std::vector<size_t>& s();
+    sample_data(int,std::vector<int>);
+    int n_phys_sites() const;
+    std::vector<int> s() const;
+    int& n_phys_sites();
+    std::vector<int>& s();
 private:
-    size_t n_phys_sites_;
-    std::vector<size_t> s_;
+    int n_phys_sites_;
+    std::vector<int> s_;
 };
 
 namespace sampling{
     template<typename cmp>
-    std::vector<sample_data> tree_sample(size_t,graph<cmp>&,size_t);
+    std::vector<sample_data> tree_sample(int,graph<cmp>&,int);
     template<typename cmp>
-    std::vector<sample_data> tree_sample(graph<cmp>&,size_t);
+    std::vector<sample_data> tree_sample(graph<cmp>&,int);
 }
 
 #endif
