+++ cpp/ttn_ops_born.hpp
@@ -8,7 +8,7 @@
 #include "sampling.hpp"
 
 template<typename cmp>
-void canonicalize(graph<cmp>&,size_t);
+void canonicalize(graph<cmp>&,int);
 template<typename cmp>
 void canonicalize(graph<cmp>&);
 // template<typename cmp>
@@ -17,8 +17,8 @@ void canonicalize(graph<cmp>&);
 template<typename cmp>
 double calc_z_born(graph<cmp>&);
 template<typename cmp>
-std::vector<double > calc_w_born(graph<cmp>&,std::vector<sample_data>&,std::vector<size_t>&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&);
+std::vector<double > calc_w_born(graph<cmp>&,std::vector<sample_data>&,std::vector<int>&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&);
 template<typename cmp>
-std::vector<double> update_cache_w_born(graph<cmp>&,size_t,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&);
+std::vector<double> update_cache_w_born(graph<cmp>&,int,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&,std::vector<std::vector<array1d<double> > >&);
 
 #endif
